- Major extension reworks  to match ares 3.0+ changes - Fixed some stuffs parsed twice and messup the output value (especially WH armor ) -significantly reduce size by eliminating `Extension` virtual functions , and add more concept to check if function is actually exist

+++ src/Misc/Ares/Hooks/Hooks.Others.cpp
@@ -2312,7 +2312,7 @@ DEFINE_OVERRIDE_HOOK(0x746B89, UnitClass_GetUIName, 8)
 	const auto& nVec = GetGunnerName(pType);
 
 	const wchar_t* Text = nullptr;
-	if (nCurWp < nVec.size())
+	if (nCurWp < (int)nVec.size())
 	{
 		Text = nVec[nCurWp].Text;
 	}
